- remove seemingly unnecessary 'componentFactory'.  It was overridden by   stylesComponentFactory. - clean up stylesComponentFactory   - add ability to drag group nodes - revise the way we ignore edge bendpoints.  Instead of copying and   modifying DefaultConnectorTerminal we define ExtendedBaseEdge,   which  overrides the bendpoints getter and setter.  We register the   custom element by registering an elementFactory. - clean up dir structure   - delete unused ShapesComponentFactory
